requirements.txt
av==11.0.0
accelerate==0.32.1
diffusers==0.24.0
einops==0.8.0
imageio-ffmpeg==0.4.9
insightface==0.7.3
omegaconf==2.2.3
opencv-python==4.9.0.80
ort_nightly_gpu
pillow==10.2.0
pyyaml==6.0.1
safetensors==0.4.3
torch==2.3.1+cu121
torchaudio==2.3.1+cu121
torchvision==0.18.1+cu121
tqdm==4.66.2
transformers==4.41.2
xformers==0.0.27
yacs==0.1.8